Shift methods to better locations
Issue #332 Shift filterBank methods to Spectrum tools because they are now generalised for all frequency scale transformations.
